compositor: various cursor fixes
Ensure that the cursor image is set again after the user switches back to the wall. Make sure that handle_button still receives button release events when the user plays an instance. Warp the cursor to the center of the screen whenever window focus changes, instead of only when pointer constraints are handled.