config, wall, layout: start implementing layout generators
Still in a very rough state. Probably plenty of large bugs left to be found and fixed. Need to add support for specifiying which instances will be hit by `wall_reset_all` and make some other tweaks with how inputs are handled in the wall module to allow for proper instance moving support.