compositor: always create and check wl_resources
I think it's technically incorrect behavior for me to not make a wl_resource when I'm given an ID to make one with? In some cases (like in xdg_decoration) creating the resource is needed so that you can immediately post an error with it. Also added allocation checks to each callsite of wl_resource_create.