server: provide wl_shm global
I hope to follow a lot of the patterns in this commit for the rest of
the globals that need to be implemented (e.g. USE_X_VERSION,
SRV_X_VERSION, wl_xxxx_g, etc.)

This is still untested. I'm not too sure about the shenanigans with
`dup` and `close`, but I think it's probably better than trying to
ref-count the original file descriptor for the SHM pool.

It's also entirely possible (likely, even?) that there's no real reason
for us to keep the FD around at all. The only thing that clients should
be realistically using wl_shm for is cursor images, and we don't care
about them.