server: provide wl_compositor global
Similarly to wl_shm, this is still untested. There are a few noteworthy
decisions in this commit.

First, state is stored inside `struct server_region`. Typically, when a
method requires a `wl_region` as a parameter, it has copy semantics
(e.g. `wl_surface.set_input_region` and `wl_surface.set_opaque_region`).
Because of this, we need to keep track of what the contents of the
region actually are.

Secondly, we do buffer all of the pending surface state. This is because
later on we will need to make use of `wl_subcompositor` to reposition
various surfaces, which requires us to call `wl_surface.commit`. We do
not want to commit some partially-submitted client state alongside our
subsurface state.

Finally, there are several methods which are left unimplemented
(alongside events such as `wl_surface.enter` and `wl_surface.leave`.) We
have a fixed, known set of clients (GLFW Wayland, Xserver's Xwayland
backend, and Mesa), so we only need to implement what they require. This
helps keep the code a bit smaller and simpler.