server: handle more allocation failures
Killing individual clients instead of the entire server in the event of
an allocation failure is not a huge improvement. However, I think
throwing asserts everywhere to check for allocation failure is not very
good.

There are a handful of cases left which I might get to fixing up at some
point (e.g. in the wl_seat handler), but some do not make much sense to
handle, such as wl_shm.format, which should only happen at startup and
cannot signal failure to anything.