server: rework surface roles
To be honest, I'm not sure if this resource destruction method is safe
(I hate libwayland.) Hopefully it is, though, because it makes the
relationships between wl_surface <-> xdg_surface <-> xdg_toplevel much
simpler to reason about.

I think using vtables like this (similarly to the buffer system) is
better than having a bunch of wl_signals that are only meant to be used
by a single consumer at a time.