server: create remote window to show UI
The intention is to show all of the user's instances via subsurfaces.
The window just serves a surface to display all of them on.

The background color will be made configurable at a later date. I may
also experiment with allowing more configuration (e.g. a background
image) at some point, but certainly not right now.