server: add server_view
This type will be used to represent surfaces which should be shown to
the user (i.e. xdg_toplevels for native Wayland clients, and X11 windows
from Xwayland clients at some point in the future.)

The xdg_toplevel implementation is not included in this commit.
Additionally, there need to be some events to wire this system up to the
rest of waywall.