server: provide wl_seat global
This one is a big one. There are definitely some weird edge cases that
are not being handled at the moment (e.g. out of bounds pointer motion
events, weird shenanigans with enter/leave, etc), but it does appear to
give input events to the correct objects.

The next steps are obviously to fix the above issues, add events for the
rest of the codebase (the wall) to receive user input, and add support
for pointer constraints and relative pointer. After that, the compositor
modules should be mostly ready to go.