server: set correct buffer implementation
Another small oversight from not having tested the code after writing
it. It's kind of annoying to unit test a compositor, though. I might try
to write some test clients at some point, but I'd be particularly
interested in trying to write a Wayland protocol fuzzer to catch the
ungodly amount of bugs that are probably lurking in this code.