server: use server_view for server_seat_g.input_focus
It makes more sense to use `struct server_view` directly, considering
only views are meant to be given input focus. This removes the need for
the annoying `get_view` method in `server_surface_role`, which would
cause problems with trying to update the input focus to a newly created
view in a `server_ui.events.view_create` handler.