server: provide zwp_pointer_constraints global
It doesn't actually do anything yet, but it responds to whatever
requests GLFW submits and that is enough to keep the game from crashing.
At this point, it is possible to boot up Minecraft within waywall, load
a world, and perform mouse inputs. (Keyboard inputs still crash.)

I plan to refactor how seat objects and input focus are handled, and
then I will make the pointer constraints implementation actually lock
the pointer when it is correct to do so.