server: move seat ownership from server_seat_g
Trying to deal with binding the seat globals in the server module only
to immediately pass them to the wl_seat module proved to be quite a
hassle. It is also not ideal to have other modules (pointer constraints
and relative pointer) to be reaching around inside the seat module.

Keyboard (`wl_keyboard`) and pointer (`wl_pointer`) objects are lazily
created to avoid their initial events being dropped. The seat_keyboard
and seat_pointer events do not instantiate them; instead, they are
created as needed when `server_get_wl_keyboard` and
`server_get_wl_pointer` are called.

The wl_seat module is still the only one which is supposed to add
listeners to the keyboard or pointer objects.