server: send placeholder keyboard settings
This is necessary to prevent the game from crashing. It does appear to
process keyboard input correctly, though.

This should conclude the 5000 lines of compositor code I needed to write
before being able to do anything else of value. There are still plenty
of loose ends to tie up, but this is good enough to get started on the
actual wall code again.

Only took 14k+ LOC and several rewrites to get to this point (: