inotify: add inotify module
Many distributions place low limits on inotify resources (instances and
watches) by default. As such, it is desirable to share one inotify
instance between all of the modules in waywall that want to receive
events about files on disk. This inotify module is meant to provide an
easy way to do that.

A linear array is used in `struct inotify` to store all of the listeners
for several reasons.

  1. It's simpler. It's just an array, not a hashtable or anything else,
     and it has fast lookup.
  2. inotify_add_watch() returns watch descriptors in sequential order
     from 1 to INT_MAX. It wraps around after reaching INT_MAX, but this
     should never happen (and there is an assert in place to catch this
     anyway.)