server: add synthetic input functions
Unfortunately, the only way to send input to Wayland surfaces is to send
them fake enter/leave events so that they think they have input focus.
This is a problem because GLFW considers a window with input focus to be
fully focused, which in turn causes SpeedrunIGT to start the timer if we
do not buffer a pause from the world generation screen.

Fortunately, the new `f3PauseOnWorldLoad` from StandardSettings means
that this is not a problem on newer versions where we would be able to
use the GLFW Wayland backend.

Whenever Xwayland support is added, I will have to investigate whether
or not it is worth adding a separate codepath to send synthetic input
events over the X11 connection.