server: add cursor module
Managing the cursor image is complicated enough that I think it deserves
its own module. I'm still not sure about keeping `server_backend` and
`server_ui` as plain structs inside of `server`, either; I might end up
switching those to be pointers as well and split the backend into its
own module.

Anyway, this commit implements the ability to load cursor themes, update
the cursor image, and hide the cursor. There will be at least one config
option for this (the cursor theme), but I may add another for the actual
cursor icon which is picked from the theme. It might also be a nice QoL
addition to allow for providing PNGs or other images.