server: create xkb_keymap for remote keymap
The keymap management logic is becoming an absolute disaster pretty
quickly, but I'd like to get the config actions system working before I
start to overhaul it.

The context should not be stored in the config struct, but instead
shared between places which need it (i.e. config and server_seat_g).
There should also probably be a custom error logger in the xkb_context.