config: make it harder to break the Lua VM
I'm not interested in trying to fully lock down everything, but my goal
is to at least make it non-trivial to crash the waywall process.

You can still hardlock it by running an infinite loop in your
configuration, but I will look into the performance impact of running
without the JIT so that debug hooks can be set to prevent that.