server: do not embed server_ui directly
There is some cost with having an additional pointer dereference
whenever some other code needs to poke around in the internals of
server_ui, but it's not the end of the world. I'd rather all of the
server module's resources be consistently heap allocated, instead of
worrying about which ones are actually just embedded inside another
struct.