server: always send local keymap
Sending the remote keymap in the event that there is no custom keymap
might be able to cause a problem? The server would think that it is
using an XKB keymap with the default options (all blank strings), while
the client would think it is using whatever the remote compositor
provides (which may use different options.)

It would be possible to pass through the remote keymap and update the
local keymap to equal it, but that is additional complexity I don't want
to deal with. Users can copy the same keymap settings that their remote
compositor is using into their waywall configuration.