config/wall: add set_resolution to lua API
I'm not confident in the correctness of the case where width or height
are greater than the window size, since the compositor is supposed to
throw a protocol error if the crop is partially outside of the surface
buffer (which should be able to happen if the game takes too long to
update itself, I think?)

The surface state logic also needs an overhaul anyway, since right now
there are a ton of needless commits and, in cases where multiple
operations happen in sequence, flickering or "not perfect" frames. And
we would never want to violate the wishes of the Wayland protocol
designers! (: