config/server: add confine_pointer option
This is nice for ensuring that you don't fling your cursor to another
monitor when you open an inventory or menu. Unfortunately, it seems like
the pointer constraints crash I discovered a while back in river still
seems to exist in some form (I thought it was fixed). That or my river
is just more out of date. Awesome