config: process layout table
This is the first of several steps for implementing layout generators.
Layout generators will be able to implement any subset of these
functions that they choose. Some will need a very minimal set (e.g.
standard wall), while others will need more (e.g. dynamic wall).

The `manual` function will be one where the user's Lua configuration can
request that a new layout is used. This removes the need for C code to
make assumptions about the user's configuration (e.g. that a layout
request should be issued as a result of resetting an instance.)

I'm not totally sure if the `manual` function is the best idea, but it
seems better to set a flag somewhere and call it after the fact rather
than giving user code the ability to provide a bunch of new layouts or
cause weird reentracy problems.