config: add instance to lua API
This function allows for getting the state of a given instance. I don't
think there's enough overhead involved in the LuaJIT <-> C barrier to
provide a function which returns the state of several (or all)
instances.

There should be more documentation about the format of the table that
this function returns, but that can happen whenever there's proper
documentation (i.e. not just luadoc comments.)