config/wall: implement custom layouts
This is not fully tested yet. The next few commits will flesh this out a
bit more (perf/API improvements, etc).

The performance of `layout_wall` and `get_hovered` can certainly be
improved quite a bit. In particular, for `layout_wall`, I would like to
make it so that only the necessary calls are made to the remote Wayland
server (i.e. diff the new layout with the previous.)

I would also like to add support for Z ordering in the future, as it
makes some interesting wall layouts more possible.

Colored rectangles (and potentially PNG sources, too) will become
available at some point in the future.