server: add UI transactions
The intent of this functionality is to allow for changing the properties
of multiple subsurfaces with one atomic operation. That being said, it
is not atomic yet (as is mentioned in `transaction_apply`) because all
of the subsurfaces are set to desynchronized mode upon creation.

I will have to do more investigation into how synchronized and
desynchronized modes operate to decide if keeping all of the subsurfaces
in synchronized mode is desirable. I think it might make sense to keep
them synchronized while on the wall and then set them to desynchronized
mode when ingame.

Despite the fact that the transaction logic is not atomic, it can still
reduce the number of needless wl_surface.commit requests that get issued
as a result from applying new layouts in wall code. I would like to
further reduce the amount of spurious Wayland traffic by only sending
messages that actually change surface state (i.e. diff against old
state).