cpu/wall: add cgroup CPU management
This is still a bit barebones, since it doesn't have preview threshold
or prioritization (for locked instances.) However, it does appear to
work. Kind of.

The current cgroups setup code doesn't do enough to allow the user to
move processes into waywall cgroups. I think the reason it used to work
was because the cgroupfs was mounted again at /sys/fs/cgroup/waywall,
which ended up functioning like another root cgroup? Or something. I'm
not entirely sure.

Either way, the user needs to have permission to move processes from the
cgroup they start from into the waywall cgroup. In systemd-based
environments, I think it would be fine to create the cgroups inside the
user slice (and it may not even need root permissions, I haven't tried
yet.) On systems that do not make use of systemd, I think the only
option is to allow the user to move all processes out of the root cgroup
(or force them to run a script to do it with root privileges.) I don't
want to force waywall to be exclusive to systemd, but it would make
things really convenient here.