cpu: use systemd user slice for cgroups
This is *still* not fully functional, unfortunately - but it's pretty
close. Using systemd's cgroups should eliminate the need for users to
enter their root password in order to setup cgroups for waywall.

The one remaining issue is that instances are usually spawned in another
cgroup which the user does not control (user@XXXX.service ->
session-X.scope). Thus, they cannot be moved into the waywall cgroup,
which is created at user@XXXX.service -> session.slice.

However, this can be resolved with the use of `systemd-run` to spawn
instances within the waywall cgroup automatically. That will just need
to be another functionality in the `waywall-launch` wrapper command,
whenever that gets created.

Non-systemd support will be implemented eventually, probably, although
I don't think there's any great way to do it that doesn't involve
egregious overuse of root.