tools: add waywall-launch
Currently, this just sets WAYLAND_DISPLAY and executes the provided
command. The next step is to use systemd-run (if available) to place the
new instance inside of the waywall cgroup. At some point in the future,
I would like to figure out how to prevent environment variables from
propagating to child processes in order to ensure that any apps spawned
by Minecraft go back out to the host session.