config/server: add input remapping
I'm sure there are plenty of edge cases to be uncovered with this
behavior, but it seems to work from a cursory test, which is a pleasant
surprise.

Remappings are non-recursive (i.e. you can remap two keys to one another
to get behavior which is legal by speedrun.com rules, and they will not
cause an infinite loop or any problems).

Some things to consider and improve upon later:
  - The configuration should probably be rejected if a remapping shadows
    an action keybind, since that's how they behave.
  - There should probably be a warning if the user's rebinds are not
    legal per speedrun.com rules (i.e. more than one input mechanism can
    trigger a certain kind of input)
  - Buttons held as a result of a remapping should be handled properly
    (although this is a general concern with the implementation of
    wl_seat). I believe held keys are currently handled correctly, at
    least.