tools/waywall-launch: do not use systemd-run
Trying to paper over the issues caused by systemd-run was interesting,
but appears to ultimately not be worth doing. I found that killing the
systemd-run process does not kill the service it started, which means
that the "Kill" button in MultiMC/PrismLauncher becomes useless. I think
it would be possible to work around this by talking to systemd over
D-Bus directly and skipping systemd-run, but that's work I don't
particularly want to do.