cpu: remove systemd cgroups support
At some point I would like to investigate this again. This setup is very
bad (exposing cgroup.procs to the current user is probably not a good
thing to be doing) and I do not like the mandatory root privileges to
setup cgroups.

It should be possible to make use of systemd, but I think it will be
necessary to talk with it over dbus manually.