main/inotify: handle IN_IGNORED events correctly
This fixes the error logs when shutting down due to `config_unsubscribe`
attempting to remove a bunch of invalidated watch descriptors.

I think IN_IGNORED still needs to be handled properly in instances,
although the only case where it should occur with wpstateout.txt is if
the user deletes it for some reason.