server: use server_cursor_config struct
I think I want to use this pattern for every object which depends on
values from the user's configuration file. Any resources which need to
be created should be stored in the `struct xxx_config`, so that it can
be either destroyed upon failure or applied without any chance of errors
upon success.