server: make server configuration atomic
It's a bit annoying to handle configuration in this manner but it does
ensure that updating configuration for every component in the server
module is an atomic operation.

The one tricky part is the transfer of ownership - when the server
configuration is applied, `server_config_destroy` should no longer
destroy any of the values that it contains, since their ownership has
been transferred to other server components.