server/wall: fix out_of_buffer error when cropping
This is far from the ideal solution (which would be to have a proper
transaction system, as noted in the new TODO comment in
transaction_apply.) This solution does result in some weird frames where
the view is stretched or squished when it shouldn't be. However, it does
work without crashing, which is the most important part.

This can still crash in the case where the view crop is not updated when
the window is sized back down.

At some point in the future it would be an interesting challenge and a
nice improvement to have a proper transaction system that waits for the
remote clients to update their surfaces before applying all of the
changes. I'm not interested in doing that right now, though.