server: fix place_above crash
wl_subsurface.place_above requires that the provided surface is either
a parent or sibling of the target. A sibling must be a subsurface (i.e.
it must be a currently visible view.)

When committing a transaction, the views included in the transaction are
enumerated from head to tail. However, since wl_list_insert places
elements at the head of the list, this ends up causing changes to be
applied in LIFO order.

The wall module would mark each view as above the previous view that it
added to the transaction, so the previous view would never have a
subsurface associated with it yet if it previously did not.