server: don't use used_source error code
Ok, this is a bit strange. For some reason, the used_source error code
does not appear to exist for me on a fresh Arch installation. It's also
not present in the Wayland XML protocol definition on my existing Arch
install, but it is present in the headers. Not really sure why.