server: listen for remote data offers
Currently, we just keep track of the MIME types associated with each
offer and do all of the necessary cleanup for expired offers.

The next step is to provide remote data offers to local clients, which
will require a decent amount of plumbing. After that, the next step will
be to provide data offers from local clients to the remote compositor,
which will probably be even more painful.