server: store last input event serial
This whole thing with using serial numbers as a rationale for making
requests is so silly. It seems that GLFW takes this approach of keeping
a recent serial number around and using it for requests like
wl_data_device.set_selection. wlroots just has some weird ring buffer
mechanism which seems to throw out some obviously invalid serials and
that's about it. Whatever.