server: provide local offers to remote compositor
This concludes the clipboard implementation (thank god). I think this
should be fully correct (I haven't found a way to crash it or trigger a
memory leak of some kind yet), but who knows. The requirement to have
wl_data_source.cancelled be sent before wl_data_device.selection is a
bit sketchy, but there's not much I can do about it.