config/wall: replace layout system with events
This should probably have been split up over multiple commits, but oh
well. The previous layout system had a few shortcomings which this new
system should hopefully address:

  1. It was not possible to update the layout generator without
     reloading the configuration.
  2. The `request_layout` system was a bit of a hack and is not very
     nice to use. Being able to directly update the current layout is
     more ergonomic.

In particular, it was going to be very difficult to implement the
benchmark functionality in a reasonable way with the old layout system.
The new system allows for replacing and restoring event listeners, which
should make it quite easy to do now.

One possible use case which this new system still fails to address is
listening for events for debugging or other purposes (I'm not sure what
reasons people would have to listen for events if not to generate
layouts, but you never know.) It is technically possible to replace the
event listener table with a blank one and then install trampolines to
the previous event handlers, but that's not great. Regardless, I think
this system is still an improvement over the previous layout system and
it should hopefully be fine.