server: allow infinite remote color buffers
The previous system had a limit of 64 separate colors at once, which is
pretty reasonable but could be reached if you wanted to with a layout
generator. Now, the main issue is with the host compositor's ability to
handle hundreds of subsurfaces.

A secondary goal of this change was to try and fix the bug where the
background surface would not update on configuration reload.
Unfortunately, it has not fixed the problem, and I still have not
managed to figure it out even with some additional debugging.