config: disable JIT and install hook by default
This is intended to prevent users from causing waywall to freeze with
broken configurations. There is an arbitrary limit of 50,000,000
executed instructions in one call from C code to Lua code, at which
point an error is thrown.

It should be noted that this does not make it completely impossible to
hardlock waywall with a broken configuration. Because the debug hook
throws an error, users can wrap their code with `pcall` to catch the
error and continue executing code. However, I think this is a reasonable
tradeoff because pcall should not really be used anyway.