server/wall: add dependency system to transactions
This is one of a few steps to try and improve the visual quality of
various state changes (mainly window resizes), although this might be
the hardest one.

Transactions now listen for changes to window size before applying the
transaction, with a hard timeout of 100 milliseconds before going
through with the transaction anyway (in the case of instances which are
lagging behind during world generation, or frozen, etc.)

Unfortunately, other sources of events (downstream Minecraft clients and
the remote compositor) can still cause surface commits which result in
imperfect frames. Additionally, window resizes still result in subpar
results due to the layout transaction occurring after a commit happens
in response to the xdg_surface configure events. These are some of the
aforementioned steps that still need to be put in place.