server: add transaction asynchronicity
This isn't actually used anywhere in the wall code yet, but this can be
used to help improve input latency (freezing during transactions) at the
expense of some potentially imperfect frames. This is a pretty
reasonable tradeoff, but I think it's cool that there's the potential to
make it an individual choice for end users.