server/wall: improve remote window resizing
The layout generator now has a chance to provide a new layout which will
be applied before the root UI surface is committed and the user sees the
window size change, which looks better.

Additionally, setting the behavior of the transaction in `layout_active`
when no specific resolution is in use to TXN_BEHAVIOR_ASYNC ensures that
we do not sit around waiting for resize events, which greatly improves
the visual quality of the window when an interactive resize is
happening.