config: allow actions to pass through input
User actions can now optionally return a boolean value to indicate
whether the input has been consumed. If they return `true`, the behavior
is the same as before. If they return `false`, the input event is passed
through to the active instance (if any).

This is desirable if keybinds may not always perform an action, so that
the associated input can be used normally.