server: provide wl_output global
Without any wl_output, GLFW will decide that enabling fullscreen ingame
should cause the window to switch back to whatever size it was initially
launched with.

Unfortunately, the behavior of fullscreen is still incorrect if the
window size changes, but this is GLFW's fault. I'll have to look into
filing a bug report or something for it.