server: do not open X11 socket for Xwayland
My launch code appears to do something subtly wrong which causes the X
server to get stuck, and I haven't been able to figure out what it is.
Fortunately, it seems that letting the X server open its own display
solves the problem.

This also removes some code, but does mean that we won't know what X11
display is in use until after the X server has launched. It shouldn't be
a huge problem to write this to /tmp/waywall-display or the environment,
though.