server: implement X11 window manager
This was extremely painful to get working. But it does appear to work
now!

Things still seem to function mostly OK despite some code still not
being present (see the TODOs in `xwayland.c`), but there are some
problems I have yet to figure out (such as XKB being unhappy with some
esoteric keyboard layouts.)

There are some other improvements which still need to be made (improved
compatibility/testing for older Xwayland versions which lack features, a
separate codepath for the `fake_input` module for X11 clients, etc), but
it's a good start.