config: load all lua libraries
Loading only a subset of the Lua standard libraries was intended as a
sandboxing mechanism, but there are plenty of valid reasons for people
to want to access packages like `os`. I don't think paranoia about
malware being distributed via config files is a very valid reason to do
this yet.