server: remove transaction behavior system
For right now, I don't want to include this. It's a neat idea for
maximum performance (i.e. minimum latency at the cost of imperfect
frames), but it's somewhat messy.

I will need to improve the code responsible for resizing the active
Minecraft instance when the waywall window is resized now, though, since
without TXN_BEHAVIOR_ASYNC it is quite unpleasant to resize the waywall
window by dragging the mouse cursor.