server: fix transaction crop bug
The fact that this ever worked is actually kind of surprising now that I
know the cause. Still have no idea why this works on other compositors
like KWin (perhaps they have bugs that allow this to not emit a protocol
error?).

The issue was that the new, larger buffer would be committed after we
already attempted to commit a source box which would be out of the
bounds of the previous, smaller buffer. The fix is to ensure that the
source box is committed with the new buffer instead of before it.

The transaction system is still a complete disaster but at least this
should fix the one known crash.