wrap: add support for floating windows
This will be used for Ninjabrain Bot support.

Currently, any number of windows can be opened in the waywall session
(only X11 windows, since there is no public Wayland socket) and any
windows after the first will be considered to be floating. The user can
click on floating windows to give them input focus.

The next steps are to allow for dragging floating windows, hiding them,
and restacking them.