config/server: add option to enable tearing
Implementing the tearing control global would require more work and
would not be desirable anyway, since Mesa only uses the tearing control
protocol in Vulkan contexts (not OpenGL). As a result, the tearing
control global would only be used by X11 Minecraft instances, which are
currently not supported anyway.

Some compositors might not enable tearing because the async presentation
hint is only set on the root surface and not the Minecraft subsurface,
but wlroots seems to do it.