config/wrap: add exec function to lua API
This function allows the user to execute commands as subprocesses of
waywall.

The `io.popen` function from the Lua standard library is not ideal for
this use case because Lua will call pclose() on any living subprocesses
when the Lua VM is destroyed. This means that waywall will freeze
whenever the user reloads their configuration while a subprocess is
still alive.