config/wrap: add option to anchor floating windows
This can be used if the user wants to anchor Ninjabrain Bot to a side or
corner of the waywall window. Anchoring a window will keep it in the
specified location.

If a corner is specified, the Ninjabrain Bot window will fill that
corner of the waywall window. If a side is specified, the Ninjabrain Bot
window will be centered on that side of the waywall window.