config, lua: add coroutine infrastructure
These changes allow waywall API functions to make use of coroutines for
pausing execution of keybind handlers until an event occurs.

I would like to expand this functionality to event listeners as well,
but the entire config subsystem needs a bit of an overhaul at this point
as it is quite complicated and the Lua interop code is extremely
fragile.